Added profile page to upload, save, and delete music
Users have a profile page that has a button to upload their music as well as add an optional cover image, title, and description. This is saved to their profile, so if users log out and log back on they can still listen to their music and view its information. There is also a button to delete their music from their profile.